WIP: restructuring application
