+++ api/server.go
@@ -1,0 +1,6 @@
+package api
+
+// type Server struct {
+// 	config util.Config
+// 	store  db.store
+// }
