+++ cmd/web/helpers.go
@@ -4,8 +4,6 @@ import (
 	"fmt"
 	"net/http"
 	"runtime/debug"
-
-	"github.com/spf13/viper"
 )
 
 // serverError: writes an error message and stack trace to errorLog,
@@ -25,35 +23,4 @@ func (app *application) clientError(w http.ResponseWriter, status int) {
 // notFound: convenience wrapper around clientError that sends a 404 Not Found response
 func (app *application) notFound(w http.ResponseWriter) {
 	app.clientError(w, http.StatusNotFound)
-}
-
-// Config stores all configuration of the application.
-// The values are read by viper from a config file or environment variable.
-type Config struct {
-	DBDriver     string `mapstructure:"DB_DRIVER"`
-	DBSource     string `mapstructure:"DB_SOURCE"`
-	MigrationURL string `mapstructure:"MIGRATION_URL"`
-}
-
-// loadConfig: loads configuration variables
-func (app *application) loadConfig(path string) (config Config) {
-	viper.AddConfigPath(path)
-	viper.SetConfigName("app")
-	viper.SetConfigType("env")
-
-	viper.AutomaticEnv()
-
-	err := viper.ReadInConfig()
-	if err != nil {
-		trace := fmt.Sprintf("%s\n%s", err.Error(), debug.Stack())
-		app.errorLog.Output(2, trace)
-		return
-	}
-	err = viper.Unmarshal(&config)
-	if err != nil {
-		trace := fmt.Sprintf("%s\n%s", err.Error(), debug.Stack())
-		app.errorLog.Output(2, trace)
-		return
-	}
-	return config
 }
